+++ docs/changelog.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug: CSS files in `convert_markdown` where not amended as inputs.
   Related to this, an extra argument was added to specify CSS files in a `plan.py`.
 - Fix bug: put header output of `markdown_katex` plugin in the HTML header.
-
+- Fix bug: rewrite paths to CSS files in `convert_markdown`
+  to be relative to the parent of the output HTML file.
 
 ## [1.4.0] - 2024-08-27 {: #v1.4.0 }